requirements.txt
Django>=2.2.0,<3.0.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.5,<2.9.0
django-heroku==0.3.1
gunicorn==20.0.4
dj-database-url==0.5.0
flake8>=3.8.1,<3.9.0